Fixed a crash involving UPNP when shutting down the gserver. Fixed a bug with CString that broke all the packets.  All of them.

+++ server/src/TPlayer.cpp
@@ -2051,7 +2051,7 @@ void TPlayer::setFlag(const CString& pFlagName, const CString& pFlagValue, bool 
 bool TPlayer::msgPLI_NULL(CString& pPacket)
 {
 	pPacket.setRead(0);
-	printf("Unknown Player Packet: %i (%s)\n", pPacket.readGUChar(), pPacket.text()+1);
+	printf("Unknown Player Packet: %u (%s)\n", (unsigned int)pPacket.readGUChar(), pPacket.text()+1);
 	for (int i = 0; i < pPacket.length(); ++i) printf("%02x ", (unsigned char)((pPacket.text())[i])); printf("\n");
 
 	// If we are getting a whole bunch of invalid packets, something went wrong.  Disconnect the player.
